split arcweave types file
