+++ Source/arcweave/Public/ArcscriptInputType.h
@@ -1,0 +1,8 @@
+
+
+#pragma once
+UENUM(BlueprintType)
+enum FArcscriptInputType {
+	CONDITION,
+	SCRIPT
+};
